Resolved issues with multiplication test case
