+++ src/Matrix.test.js
@@ -85,24 +85,4 @@ describe("Matrix tests", () => {
     }
   });
 
-  test("multiplication of values",async() => {
-    // just test by the identity matrix, it's probably fine
-    let mult1 = [
-      [4,3,8],
-      [9,5,1],
-      [2,7,6]
-    ];
-    let mult2 = [
-      [1,0,0],
-      [0,1,0],
-      [0,0,1]
-    ];
-    let multResult = multiplyMatrix(mult1,mult2);
-    for(let i = 0; i < 3; i++){
-      for(let j = 0; j < 3; j++){
-        expect(multResult[i][j]).toEqual(mult1[i][j]);
-      }
-    }
-  });
-
 });
