+++ src/calculate.test.js
@@ -2,7 +2,7 @@
 /* eslint-disable testing-library/no-unnecessary-act */
 /* eslint-disable testing-library/no-node-access */
 import { createRoot } from "react-dom/client";
-import { addMatrix } from "./calculate";
+import { addMatrix, multiplyMatrix } from "./calculate";
 
 describe("Calculate tests", () => {
   let container, root;
@@ -30,4 +30,23 @@ describe("Calculate tests", () => {
     }
   });
 
+  test("multiplication of values",async() => {
+    // just test by the identity matrix, it's probably fine
+    let mult1 = [
+      [4,3,8],
+      [9,5,1],
+      [2,7,6]
+    ];
+    let mult2 = [
+      [1,0,0],
+      [0,1,0],
+      [0,0,1]
+    ];
+    let multResult = multiplyMatrix(mult1,mult2);
+    for(let i = 0; i < 3; i++){
+      for(let j = 0; j < 3; j++){
+        expect(multResult[i][j]).toEqual(mult1[i][j]);
+      }
+    }
+  });
 });
